Implement regrading & unify scoring calculations
No more partial database updates*; scores are now always
going to be consistent. Not that they weren't before, but
it was getting hard to manage state.

Because a re-grade can cause more than one successful
submission to appear on the same problem+user, dropped
the unique index blocking that. Scoring logic is now robust
enough to be able to pick the earliest successful submission
using a distinct+order-by clause.

* Except one where we store penalties for leaderboards, but
that's re-computed in final calculations.